Decrease timeout and tries on check_zone_auth.
Apparantly, the c.ip6-servers.arpa have decided they do not want to reply to our nsmaster anymore. The timeout and associated retries was making it take 25 seconds per try before a new server was selected, die to the timeout being 5 and tries being 5 as well. Decreased this for this particular referal to a timeout of 3 seconds and no more than 2 tries. Fixes the amount of false positives in our monitoring.